Add connector context manager to libs
